LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/top.meta/origin-when-crossorigin/img-tag.http-expected.txt

CONSOLE MESSAGE: Failed to set referrer policy: The value 'origin-when-crossorigin' is not one of 'no-referrer', 'no-referrer-when-downgrade', 'same-origin', 'origin', 'strict-origin', 'origin-when-cross-origin', 'strict-origin-when-cross-origin' or 'unsafe-url'.
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?redirection=keep-origin&action=purge&key=c3d07094-f231-4576-8ccc-e7f10024fe55&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?redirection=no-redirect&action=purge&key=59f12156-ab76-4d44-9f87-a60deb4ffb69&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?redirection=swap-origin&action=purge&key=6e58c225-f459-4144-97cf-fa1d5d4bfe2d&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/image.py?redirection=keep-origin&action=purge&key=3d9fb5e4-358f-4127-82a5-3e9dece857b4&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/image.py?redirection=no-redirect&action=purge&key=cb40303b-45db-4b6c-8574-d68cbe8bf349&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/image.py?redirection=swap-origin&action=purge&key=1a08ac8b-54be-4d8f-8e22-950c52c9ad9e&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/image.py?action=purge&key=38a2c766-c2c8-4615-bfef-d835aad986a3&path=%2Fmixed-content


Harness Error (TIMEOUT), message = null

FAIL Referrer Policy: Expects origin for img-tag to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects origin for img-tag to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects origin for img-tag to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects origin for img-tag to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects origin for img-tag to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
FAIL Referrer Policy: Expects origin for img-tag to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: object "[object Event]"
TIMEOUT Referrer Policy: Expects origin for img-tag to same-http origin and swap-origin redirection from http context. Test timed out
NOTRUN Referrer Policy: Expects origin for img-tag to same-https origin and keep-origin redirection from http context.
NOTRUN Referrer Policy: Expects origin for img-tag to same-https origin and no-redirect redirection from http context.
NOTRUN Referrer Policy: Expects origin for img-tag to same-https origin and swap-origin redirection from http context.
NOTRUN Referrer Policy: Expects stripped-referrer for img-tag to same-http origin and keep-origin redirection from http context.
NOTRUN Referrer Policy: Expects stripped-referrer for img-tag to same-http origin and no-redirect redirection from http context.
